Reduce sync events
Provide the meta that triggers a sync event so we can skip sending
extraneous sync events to the players that trigger them. Now we need
to provide a little more context in the `game:joined` event and add a
`game:update` event to help separate self-triggered events from sync events.