Add action to unimprove groups evenly
In a game of monopoly, there could be a situation where you want/need
to unimprove a hotel, but there are less than four houses available to
exchange. You cannot take less than four because you need to
unimprove evenly, and you cannot unimprove other hotels because there
are not enough houses.

In these situations you would take all available houses and distribute
them evenly between your properties; selling the difference between
buildings back to the bank (even though the bank does not receive any
houses in return).

To facilitate this, the `unimproveGroup` group action does just that
and weighs its distribution of houses to the more valuable properties
in the group. If the group does not need houses, or if there are more
than enough houses, this action will just unimprove the _least valuable_
property in the group, with the _most houses_.