Refactor server
Redux was removed in favor of a more simple system. Instead of having to keep a
redux store in sync with a database, and having various actions, action types,
and reducers to update the state; the state is now defined as a series of
single-use reducer functions.

Furthermore, instead of having the game room be a loose connection between
websockets, the redux store, and the database (along with other room-specific
jobs); it is now been expanded into multiple classes, each with their own
responsibilities.

The game room manager brokers connections between websockets (as players) and
game rooms; as well as having methods for loading and saving games to a
database. Without a redux store, game room events simply load the game from
storage, apply the appropriate reducer functions, and save the game back into
storage; sending any relevant events to players (including success or error
events back to the player responsible for triggering the event).