Use timestamps to determine newer event data
Notices did this so the toaster could determine when to actually show a toast or
if it had already been shown before. The API context can use timestamps to
prioritize merging new data into old data in the case where some events are
recieved out of order. The pre-existing notice timestamps have been removed in
favor of the new state timestamp.